[ADD]파일 삭제 로직 추가

+++ src/main/java/com/swan/picturerepository/controller/UploadController.java
@@ -5,7 +5,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
-import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
 
 import org.apache.commons.io.FilenameUtils;
@@ -24,21 +23,18 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.view.RedirectView;
 
+import com.swan.picturerepository.dto.UserFileInfoDTO;
 import com.swan.picturerepository.service.BulletinboardService;
+import com.swan.picturerepository.service.FileSearchService;
 import com.swan.picturerepository.service.FileUploadService;
 
 @Controller
 @RequestMapping("/bulletinboards")
 public class UploadController {
 	
-	@Resource(name = "imageUploadPath")
-	private String imageUploadPath;
-	@Resource(name = "thumbnailUploadPath")
-	private String thumbnailUploadPath;
-	@Resource(name = "fullUploadPath")
-	private String fullUploadPath;
 	@Autowired BulletinboardService bulletinboardService;
 	@Autowired FileUploadService fileUploadService;
+	@Autowired FileSearchService fileSearchService;
 	
 	@PostMapping(value = "/newbulletinboard")
 	public ModelAndView uploadFormMulti(Model model, HttpServletRequest req, @RequestParam("file") List<MultipartFile> fileList) throws Exception {
@@ -80,9 +76,9 @@ public class UploadController {
 				String strFileName = fileList.get(i).getOriginalFilename();
 				String strFileExtension = FilenameUtils.getExtension(strFileName);
 				String strFileId = String.format("%s.%s",uid.toString(),strFileExtension);
-				File imageFile = fileUploadService.uploadImageFile(imageUploadPath, strFileId, fileList.get(i).getBytes(), strFileName);
-				fileUploadService.uploadThumbnailFile(thumbnailUploadPath, strFileId, imageFile, strFileName);
-				fileUploadService.uploadFullThumbnailFile(fullUploadPath, strFileId, imageFile, strFileName);
+				File imageFile = fileUploadService.uploadImageFile(strFileId, fileList.get(i).getBytes(), strFileName);
+				fileUploadService.uploadThumbnailFile(strFileId, imageFile, strFileName);
+				fileUploadService.uploadFullThumbnailFile(strFileId, imageFile, strFileName);
 				userFileInfoList.add(i,strFileId);		
 			
 				sb.append(strFileName).append(",");
@@ -108,24 +104,32 @@ public class UploadController {
 
 		return mv;
 	}
-	//PUT METHOD가 잘 동작하지 않기에 POST로 처리
+
 	@PutMapping(value = "/newbulletinboard/{bulletinId}", produces = {MediaType.APPLICATION_JSON_VALUE})
 	public ModelAndView updateBulletinboard(Model model, HttpServletRequest req, @PathVariable("bulletinId") String strbulletinId) throws Exception {
 		ModelAndView mv = new ModelAndView();
 		mv.setViewName(".notdynamicjs/board/imageView");
 		return mv;
 	}
-	//DELETE METHOD가 잘 동작하지 않기에 POST로 처리
-	//@RequestMapping(value = "/newbulletinboard/{bulletinId}/delete", method = RequestMethod.POST)
+	
 	@DeleteMapping(value = "/newbulletinboard/{bulletinId}")
 	public ModelAndView deleteBulletinboard(Model model, HttpServletRequest req, @PathVariable("bulletinId") String strBulletinId) throws Exception {
-		
 		ModelAndView mv = new ModelAndView();
 		RedirectView redirectView = new RedirectView();
+		List<UserFileInfoDTO> fileList = null;	
 		
+		fileList = fileSearchService.getSearchFileListByFileId(strBulletinId);
+		if(!fileUploadService.deleteImageFile(fileList) || !fileUploadService.deleteThumbnailFile(fileList) || !fileUploadService.deleteFullThumbnailFile(fileList)) {
+			model.addAttribute("deleteErrorMsg", "파일 삭제에서 에러가 발생했습니다.");
+			redirectView.setUrl(req.getContextPath()+"/bulletinboards/" + strBulletinId);
+			mv.setView(redirectView);
+			return mv;
+		}
+	
 		if(!bulletinboardService.deleteBulletinboard(strBulletinId)) {
-			model.addAttribute("deleteErrorMsg", "삭제에서 에러가 발생했습니다.");
-			mv.setViewName("board/imageFileUpload");
+			model.addAttribute("deleteErrorMsg", "데이터 삭제에서 에러가 발생했습니다.");
+			redirectView.setUrl(req.getContextPath()+"/bulletinboards/" + strBulletinId);
+			mv.setView(redirectView);
 			return mv;
 		}
